+++ src/Qubic.vcxproj
@@ -62,19 +62,9 @@
     <ClInclude Include="contract_core\execution_time_accumulator.h" />
     <ClInclude Include="contract_core\ipo.h" />
     <ClInclude Include="contract_core\pre_qpi_def.h" />
-    <ClInclude Include="contract_core\qpi_asset_impl.h" />
-    <ClInclude Include="contract_core\qpi_collection_impl.h" />
-    <ClInclude Include="contract_core\qpi_ipo_impl.h" />
-    <ClInclude Include="contract_core\qpi_mining_impl.h" />
-    <ClInclude Include="contract_core\qpi_oracle_impl.h" />
-    <ClInclude Include="contract_core\qpi_spectrum_impl.h" />
-    <ClInclude Include="contract_core\qpi_system_impl.h" />
-    <ClInclude Include="contract_core\qpi_hash_map_impl.h" />
-    <ClInclude Include="contract_core\qpi_linked_list_impl.h" />
-    <ClInclude Include="contract_core\qpi_ticking_impl.h" />
-    <ClInclude Include="contract_core\qpi_trivial_impl.h" />
+    <ClInclude Include="contract_core\qpi_assets_impl.h" />
     <ClInclude Include="contract_core\stack_buffer.h" />
-    <ClInclude Include="contract_core\qpi_proposal_voting.h" />
+    <ClInclude Include="contract_core\qpi_proposals_impl.h" />
     <ClInclude Include="files\files.h" />
     <ClInclude Include="logging\logging.h" />
     <ClInclude Include="logging\net_msg_impl.h" />
@@ -135,6 +125,18 @@
     <ClInclude Include="platform\time_stamp_counter.h" />
     <ClInclude Include="platform\global_var.h" />
     <ClInclude Include="platform\virtual_memory.h" />
+    <ClInclude Include="qpi\impl\qpi_assets_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_collection_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_hash_map_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_ipo_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_linked_list_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_mining_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_oracle_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_proposals_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_spectrum_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_system_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_ticking_impl.h" />
+    <ClInclude Include="qpi\impl\qpi_trivial_impl.h" />
     <ClInclude Include="qpi\qpi.h" />
     <ClInclude Include="qpi\qpi_assets.h" />
     <ClInclude Include="qpi\qpi_containers.h" />
