--- src/contract_core/qpi_system_impl.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include "qpi/qpi.h"
-#include "system.h"
-
-unsigned short QPI::QpiContextFunctionCall::epoch() const
-{
-    return system.epoch;
-}
-
-unsigned int QPI::QpiContextFunctionCall::tick() const
-{
-    return system.tick;
-}
