+++ src/qpi/impl/qpi_mining_impl.h
