+++ src/qpi/impl/qpi_system_impl.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "qpi/qpi.h"
+#include "system.h"
+
+unsigned short QPI::QpiContextFunctionCall::epoch() const
+{
+    return system.epoch;
+}
+
+unsigned int QPI::QpiContextFunctionCall::tick() const
+{
+    return system.tick;
+}
