+++ test/assets.cpp
@@ -9,8 +9,8 @@
 
 #include "assets/assets.h"
 #include "contract_core/contract_exec.h"
-#include "contract_core/qpi_spectrum_impl.h"
-#include "contract_core/qpi_asset_impl.h"
+#include "qpi/impl/qpi_spectrum_impl.h"
+#include "qpi/impl/qpi_assets_impl.h"
