+++ test/qpi_linked_list.cpp
@@ -2,10 +2,10 @@
 
 #include "gtest/gtest.h"
 
-#include "../src/contract_core/pre_qpi_def.h"
-#include "../src/qpi/qpi.h"
-#include "../src/common_buffers.h"
-#include "../src/contract_core/qpi_linked_list_impl.h"
+#include "contract_core/pre_qpi_def.h"
+#include "qpi/qpi.h"
+#include "common_buffers.h"
+#include "qpi/impl/qpi_linked_list_impl.h"
 #include <vector>
 #include <random>
